requirements.txt
requests==2.31.0
StrEnum==0.4.15
python-dotenv==1.0.1
flake8==7.0.0
pytest==7.4.4
